tweet_generator/dev-requirements.txt
paho-mqtt==1.2
psutil>=5.1.2
confluent-kafka>=0.9.2
paramiko>=2.1.2
fabric>=1.13.1
